pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["Esther2002-soon <esthersoon2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.10.4"
fastapi = "^0.95.1"
sqlalchemy = "^2.0.12"
psycopg2-binary = "^2.9.6"
flask = "^2.3.2"
pymysql = "^1.0.3"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
pyjwt = "^2.7.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
